Update CNAB airgap example
This updates the airgap example to demonstrate how to include referenced
images in the bundle when it is archived to a tgz, and how to update
your manifests to use the new image location after the images are
published on the other side of the airgap.

With helm's set command this is a bit easier to accomplish, here is an
example bundle that uses helm:

https://github.com/getporter/porter/tree/main/examples/airgap

I am using yq to help modify the manifest. I think kustomize can do it
as well.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>